requirements.txt
alpaca-py==0.5.4
certifi==2022.6.15
charset-normalizer==2.1.1
idna==3.3
msgpack==1.0.4
numpy==1.23.2
pandas==1.3.5
pydantic==1.10.2
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
schedule==1.1.0
six==1.16.0
typing_extensions==4.3.0
urllib3==1.26.12
websockets==10.3
